Rename qobj to operator
