+++ qiskit/providers/aer/pulse/system_models/string_model_parser/operator_from_string.py
@@ -15,7 +15,7 @@
 
 """Module for creating quantum operators."""
 
-from . import qobj_generators
+from . import operator_generators
 
 
 def gen_oper(opname, index, h_osc, h_qub, states=None):
@@ -41,23 +41,23 @@ def gen_oper(opname, index, h_osc, h_qub, states=None):
 
         if opname in ['X', 'Y', 'Z'] and dim > 2:
             if opname == 'X':
-                opr_tmp = qobj_generators.get_oper('A', dim) + qobj_generators.get_oper('C', dim)
+                opr_tmp = operator_generators.get_oper('A', dim) + operator_generators.get_oper('C', dim)
             elif opname == 'Y':
-                opr_tmp = (-1j * qobj_generators.get_oper('A', dim) +
-                           1j * qobj_generators.get_oper('C', dim))
+                opr_tmp = (-1j * operator_generators.get_oper('A', dim) +
+                           1j * operator_generators.get_oper('C', dim))
             else:
-                opr_tmp = (qobj_generators.get_oper('I', dim) -
-                           2 * qobj_generators.get_oper('N', dim))
+                opr_tmp = (operator_generators.get_oper('I', dim) -
+                           2 * operator_generators.get_oper('N', dim))
 
     else:
         is_qubit = False
         dim = h_osc.get(index, 5)
 
     if opname == 'P':
-        opr_tmp = qobj_generators.get_oper(opname, dim, states)
+        opr_tmp = operator_generators.get_oper(opname, dim, states)
     else:
         if opr_tmp is None:
-            opr_tmp = qobj_generators.get_oper(opname, dim)
+            opr_tmp = operator_generators.get_oper(opname, dim)
 
     # reverse sort by index
     rev_h_osc = sorted(h_osc.items(), key=lambda x: x[0])[::-1]
@@ -69,11 +69,11 @@ def gen_oper(opname, index, h_osc, h_qub, states=None):
         if ii == index and not is_qubit:
             opers.append(opr_tmp)
         else:
-            opers.append(qobj_generators.qeye(dd))
+            opers.append(operator_generators.qeye(dd))
     for ii, dd in rev_h_qub:
         if ii == index and is_qubit:
             opers.append(opr_tmp)
         else:
-            opers.append(qobj_generators.qeye(dd))
+            opers.append(operator_generators.qeye(dd))
 
-    return qobj_generators.tensor(opers)
+    return operator_generators.tensor(opers)
