+++ qiskit/providers/aer/pulse/system_models/string_model_parser/operator_generators.py
