+++ test/terra/pulse/test_duffing_model_generators.py
@@ -19,7 +19,7 @@ from test.terra.common import QiskitAerTestCase
 from qiskit.providers.aer.pulse.system_models.pulse_system_model import PulseSystemModel
 from qiskit.providers.aer.pulse.system_models.hamiltonian_model import HamiltonianModel
 from qiskit.providers.aer.pulse.system_models import duffing_model_generators as model_gen
-from qiskit.providers.aer.pulse.system_models.string_model_parser.qobj_generators import get_oper
+from qiskit.providers.aer.pulse.system_models.string_model_parser.operator_generators import get_oper
 from qiskit.providers.models.backendconfiguration import UchannelLO
 from qiskit.quantum_info.operators.operator import Operator
