+++ test/terra/pulse/test_operator_from_string.py
@@ -1,11 +1,11 @@
-"""tests for qobj_from_string.py"""
+"""tests for operator_from_string.py"""
 
 # from test.terra.common import QiskitAerTestCase
 #
 # from qiskit.providers.aer.pulse.system_models.string_model_parser.qobj_from_string import gen_oper
 #
 # class TestOperatorFromStrings(QiskitAerTestCase):
-#     """Tests gen_oper in qobj_from_string.py"""
+#     """Tests gen_oper in operator_from_string.py"""
 #
 #     def test_X(self):
 #         gen_oper('X', , None, None, states=None)
